requirements.txt
streamlit
openai
python-dotenv
pandas
matplotlib
numpy
streamlit-option-menu